Updates folders structure and links to bootstrap JS files -Deletes simple HTML front-end folder -Adds folders for MVC model: controllers, models, views -Relocates files for index page to MVC folders -Adds local Bootstrap bundles JS file -Updates link to Bootstrap JS
