+++ biglibrary/big_list.php
@@ -2,15 +2,15 @@
 ob_start();
 session_start();
 
-//If user not logged in, redirect to index.php
+//If user not logged in, redirect to index_boot.php
 if(!isset($_SESSION['user'])) {
-	header('Location: index.php');
+	header('Location: controllers/index_controller.php');
 	exit;
 }
 
 require_once 'db_connect.php';
 
-$sql  = "SELECT title, publish_date, creators.first_name AS first_name, creators.last_name AS last_name, fk_user_id
+$sql  = "SELECT title, publish_date, creators.first_name AS first_name, creators.last_name AS last_name, media.img_link AS img_link, fk_user_id
 				 FROM media
 				 LEFT JOIN creators
 				 		ON media.fk_creator_id = creators.creator_id
@@ -38,39 +38,56 @@ $mysqli->close();
 <html>
 <head>
 	<title>Big List</title>
+	<meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+	<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+	<!-- <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css" integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB" crossorigin="anonymous"> -->
+	<link rel="stylesheet" type="text/css" href="css/bootstrap.min.css">
+	<link rel="stylesheet" type="text/css" href="css/style.css">
 </head>
 <body>
-	<a href="log_out.php?logout" title=""><button type="button" name="log_out">Log out</button></a>
+	<header>
+		<nav class="navbar navbar-expand-lg navbar-light bg-light">
+		  <a class="navbar-brand" href="controllers/index_controller.php">Big Library</a>
+		  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
+		    <span class="navbar-toggler-icon"></span>
+		  </button>
+		  <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
+		    <div class="navbar-nav">
+		      <a class="nav-item nav-link active" href="big_list.php">Big List<span class="sr-only">(current)</span></a>
+		      <a class="nav-item nav-link" href="log_out.php?logout">Log Out</a>
+		      <!--<a class="nav-item nav-link disabled" href="#">Disabled</a>-->
+		    </div>
+		  </div>
+		</nav>
+	</header>
 	<table>
-		<caption style="font-size: 2em">Big List</caption>
-		<thead>
-			<tr>
-				<th>Title</th>
-				<th>Author</th>
-				<th>Year</th>
-				<th>Availability</th>
-			</tr>
-		</thead>
-		<tbody>
-			<?php
-			//Show all the media in the library
-			foreach ($media as $medium) {?>
-				<tr>
-					<td><?php echo $medium['title'] ?></td>
-					<td><?php echo $medium['first_name'] . " " . $medium['last_name'] ?></td>
-					<td><?php echo $medium['publish_date'] ?></td>
-					<td><?php if(is_null($medium['fk_user_id'])) {
-											echo 'Available';
-											} else {
-												echo 'Booked';
-											}
-							?></td>
-				</tr>
-			<?php
-			}
-			?>
-			
-		</tbody>
-	</table>
+		<main>
+				<div class="container">
+					<div class="row">
+						<?php
+						//Show all the media in the library
+						foreach ($media as $medium) {?>
+							<div class="card" style="width: 18rem;">
+						  <img class="card-img-top" src="<?php echo $medium['img_link']; ?>" alt="Card image cap">
+						  <div class="card-body">
+						    <h5 class="card-title"><?php echo $medium['title']; ?></h5>
+						    <p class="card-text"><?php echo $medium['first_name'] . " " . $medium['last_name']; ?></p>
+						  </div>
+						</div>
+						<?php
+						}
+						?>	
+					</div>
+				</div>
+				
+			</tbody>
+		</table>
+	</main>
+	<!-- <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
+	<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/js/bootstrap.min.js" integrity="sha384-smHYKdLADwkXOn1EmN1qk/HfnUcbVRZyYmZ4qpPea6sjB/pTJ0euyQp0Mk8ck+5T" crossorigin="anonymous"></script> -->
+	<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
+	<script src="js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
