+++ biglibrary/db_connect.php
@@ -8,8 +8,7 @@ define('DBNAME', 'cr10_ivan_zykov_biglibrary');
 $mysqli = new mysqli(DBHOST, DBUSER, DBPASS, DBNAME);
 
 if ($mysqli->connect_error) {
-	die('Connection failed: ' . $conne_errno. ': ' . $mysqli->connect_error);
+	die('Connection failed: ' . $mysqli->connect_errno. ': ' . $mysqli->connect_error);
 } else {
 	//echo "Connection live!<br>";
 }
-?>
