+++ biglibrary/log_out.php
@@ -3,7 +3,7 @@ ob_start();
 session_start();
 
 if(!isset($_SESSION['user'])) {
-	header("Location: index.php");
+	header("Location: controllers/index_controller.php");
 } else {
 	header("Location: big_list.php");
 }
@@ -12,7 +12,7 @@ if(isset($_GET['logout'])) {
 	unset($_SESSION['user']);
 	session_unset();
 	session_destroy();
-	header("Location: index.php");
+	header("Location: controllers/index_controller.php");
 }
 exit;
 ?>
